Graphs practice
Graphs provide the ultimate in data structure flexibility. A graph consists of a set of nodes, and a set of edges where an edge connects two nodes. Trees and lists can be viewed as special cases of graphs.Graph Data Structure Mathematical graphs can be represented in data structure. We can represent a graph using an array of vertices and a two-dimensional array of edges.